1.修复解析JT808头部数据包bug 2.去掉对象池，使用缓存的方式获取对应序列化器 3.增加消息Id处理工厂接口及默认实现（设备号参数用来到时候解决Id冲突问题） 4.修改位置信息附加Id所有消息测试及对应测试报告 5.修改全局配置方法的使用

+++ src/JT808.Protocol/Formatters/MessageBodyFormatters/JT808_0x0001_Formatter.cs
@@ -23,7 +23,7 @@ namespace JT808.Protocol.Formatters.MessageBodyFormatters
         public int Serialize(ref byte[] bytes, int offset, JT808_0x0001 value)
         {
             offset += JT808BinaryExtensions.WriteUInt16Little(bytes, offset, value.MsgNum);
-            offset += JT808BinaryExtensions.WriteUInt16Little(bytes, offset, (ushort)value.MsgId);
+            offset += JT808BinaryExtensions.WriteUInt16Little(bytes, offset, value.MsgId);
             offset += JT808BinaryExtensions.WriteByteLittle(bytes, offset, (byte)value.JT808TerminalResult);
             return offset;
         }
